requirements.txt
Flask
Flask-WTF
WTForms
email-validator
mysql-connector-python
Werkzeug
python-dotenv